app_deploy/requirements.txt
fastapi
uvicorn
python-dotenv
line-bot-sdk
requests
mangum

